Dockerise project
Using instructions from this article
[link](https://nodejs.org/en/docs/guides/nodejs-docker-webapp/)

A `Dockerfile` was created very similar to the above. It was tested
with the command:

``docker build -t ngf-server .``

Once this completed, it could be ran as follows:

``docker run -p 8808:8808 ngf-server -d``

Then, we can stop this with

``docker stop $(docker ps -a | grep ngf-server)``

It was verified that this had what we required by pointing a web
browser to it and going to the /api/events URI where all the
events where returned.

The Dockerfile probably uses the latest of everything in terms
of dependencies. This is in contrast to where it was tested manually
on MS Windows 10 where Node is version 8.11.2.

    --- Lionel Saliou, Ph.D